refactor: entity 불필요 설정 제거
