+++ src/main/java/soongsil/kidbean/server/global/config/ConfigurationPropsConfig.java
@@ -2,7 +2,7 @@ package soongsil.kidbean.server.global.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
-import soongsil.kidbean.server.auth.jwt.JwtProperties;
+import soongsil.kidbean.server.auth.application.jwt.JwtProperties;
 
 @Configuration
 @EnableConfigurationProperties(value = JwtProperties.class)
