+++ src/main/java/soongsil/kidbean/server/member/domain/Member.java
@@ -5,7 +5,6 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import soongsil.kidbean.server.global.domain.BaseTimeEntity;
 import soongsil.kidbean.server.member.domain.type.OAuthType;
 import soongsil.kidbean.server.member.domain.type.Role;
@@ -14,7 +13,6 @@ import soongsil.kidbean.server.member.domain.type.Gender;
 import java.time.LocalDate;
 import soongsil.kidbean.server.quiz.domain.type.Level;
 
-@EnableJpaAuditing
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
