+++ src/main/java/soongsil/kidbean/server/quiz/domain/AnswerQuizSolved.java
@@ -5,13 +5,11 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import soongsil.kidbean.server.global.domain.BaseTimeEntity;
 import soongsil.kidbean.server.global.domain.S3Info;
 import soongsil.kidbean.server.member.domain.Member;
 
 @Getter
-@EntityListeners(AuditingEntityListener.class)
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class AnswerQuizSolved extends BaseTimeEntity {
