+++ src/test/java/soongsil/kidbean/server/global/application/config/CommonControllerTest.java
@@ -5,7 +5,7 @@ import static soongsil.kidbean.server.member.fixture.MemberFixture.MEMBER1;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
-import soongsil.kidbean.server.auth.jwt.JwtTokenProvider;
+import soongsil.kidbean.server.auth.application.jwt.JwtTokenProvider;
 import soongsil.kidbean.server.auth.util.AuthenticationUtil;
 
 @Import(TestSecurityConfig.class)
